+++ diamondback/models/GaussianMixtureModel.py
@@ -45,8 +45,8 @@ class GaussianMixtureModel( object ) :
 
     @index.setter
     def index( self, index ) :
-        if ( index <= 0 ) :
-            raise ValueError( f'Index = {index} Expected Index > 0' )
+        if ( index < 1 ) :
+            raise ValueError( f'Index = {index} Expected Index in [ 1, inf )' )
         self._index = index
 
     @property
@@ -60,7 +60,7 @@ class GaussianMixtureModel( object ) :
     @regularize.setter
     def regularize( self, regularize : float ) :
         if ( regularize < 0.0 ) :
-            raise ValueError( f'Regularize = {regularize} Expected Regularize >= 0.0' )
+            raise ValueError( f'Regularize = {regularize} Expected Regularize in [ 0.0, inf )' )
         self._regularize = regularize
 
     @property
@@ -77,12 +77,12 @@ class GaussianMixtureModel( object ) :
                 regularize : float - regularize.
         """
 
-        if ( order <= 0 ) :
-            raise ValueError( f'Order = {order} Expected Order > 0' )
-        if ( index <= 0 ) :
-            raise ValueError( f'Index = {index} Expected Index > 0' )
+        if ( order < 1 ) :
+            raise ValueError( f'Order = {order} Expected Order in [ 1, inf )' )
+        if ( index < 1 ) :
+            raise ValueError( f'Index = {index} Expected Index in [ 1, inf )' )
         if ( regularize < 0.0 ) :
-            raise ValueError( f'Regularize = {regularize} Expected Regularize >= 0.0' )
+            raise ValueError( f'Regularize = {regularize} Expected Regularize in [ 0.0, inf )' )
         self._data = [ ]
         self._index = index
         self._order = order